+++ R/textlinedetector.R
@@ -221,3 +221,4 @@ ocv_points <- function(image, type = c("topleft", "topright", "bottomleft", "bot
 
 
 
+